fix: Markdownコンテンツネゴシエーションをworker層リライトで実装し直した
React RouterのUIルートではloaderから`throw new Response(...)`しても
2xxレスポンスがそのまま返らず、Acceptヘッダー判定が実質無効化されていた。
そのためworker.tsで`/`と`/archives/{id}`へのAccept:text/markdown付きリクエストを
それぞれ`/index.md`と`/archives/{id}.md`の各リソースルートへ内部リライトする
構成に変更した。

- worker.ts: rewriteForMarkdown でリクエストを書き換える
- index[.]md.tsx: ホームページ用リソースルートを新設
- markdown.server.ts: renderHomeMarkdown を共通化して移動
- _layout._index.tsx / _layout.archives.$postId.tsx: 効いていなかった
  Accept判定を削除
