+++ app/modules/markdown.server.ts
@@ -58,9 +58,42 @@ export function renderArticleMarkdown(data: ArchiveDataEntry): string {
   return lines.join('\n');
 }
 
-export function wantsMarkdown(request: Request): boolean {
-  const accept = request.headers.get('accept') ?? '';
-  return /\btext\/markdown\b/i.test(accept);
+export function renderHomeMarkdown(data: {
+  latest: { postId: number; postTitle: string; postDateGmt: Date }[];
+  voted: { postId: number; postTitle: string; countLikes: number }[];
+}): string {
+  const lines: string[] = [];
+  lines.push('# 健常者エミュレータ事例集');
+  lines.push('');
+  lines.push(
+    '社会生活やコミュニケーションに関する暗黙知を言語化・集積し、知識のギャップを集団で補うためのプラットフォーム。',
+  );
+  lines.push('');
+  lines.push('- サイト: https://healthy-person-emulator.org');
+  lines.push('- フィード: https://healthy-person-emulator.org/feed.xml');
+  lines.push('- サイトマップ: https://healthy-person-emulator.org/sitemap.xml');
+  lines.push(
+    '- 各記事のMarkdown版: `https://healthy-person-emulator.org/archives/{postId}.md` または `Accept: text/markdown` ヘッダー付きで記事URLにリクエスト',
+  );
+  lines.push('');
+  lines.push('## 最新の投稿');
+  lines.push('');
+  for (const p of data.latest) {
+    const date = new Date(p.postDateGmt).toISOString().slice(0, 10);
+    lines.push(
+      `- [${p.postTitle}](https://healthy-person-emulator.org/archives/${p.postId}) — ${date}`,
+    );
+  }
+  lines.push('');
+  lines.push('## 最近いいねされた投稿');
+  lines.push('');
+  for (const p of data.voted) {
+    lines.push(
+      `- [${p.postTitle}](https://healthy-person-emulator.org/archives/${p.postId}) — 👍 ${p.countLikes}`,
+    );
+  }
+  lines.push('');
+  return lines.join('\n');
 }
 
 export function markdownResponse(body: string, cacheMaxAgeSec = 300): Response {
