+++ app/routes/_layout.archives.$postId.tsx
@@ -46,7 +46,6 @@ import { setIsLoginModalOpenAtom } from '~/stores/loginmodal';
 import { useSetAtom } from 'jotai';
 import { setVisitorCookieData } from '~/modules/visitor.server';
 import { Bookmark } from 'lucide-react';
-import { markdownResponse, renderArticleMarkdown, wantsMarkdown } from '~/modules/markdown.server';
 
 export const commentVoteSchema = z.object({
   commentId: z.number(),
@@ -63,11 +62,6 @@ export async function loader({ request }: LoaderFunctionArgs) {
   const url = new URL(request.url);
   const postId = Number(url.pathname.split('/')[2]);
   const data = await ArchiveDataEntry.getData(postId);
-
-  if (wantsMarkdown(request)) {
-    throw markdownResponse(renderArticleMarkdown(data));
-  }
-
   const { likedPages, dislikedPages, likedComments, dislikedComments } =
     await getUserActivityData(request);
   const isAuthenticated = await getAuthenticatedUser(request);
